--- a/src/checkqueue.h
+++ b/src/checkqueue.h
@@ -16,7 +16,7 @@
 template <typename T>
 class CCheckQueueControl;
 
-/** 
+/**
  * Queue for verifications that have to be performed.
   * The verifications are represented by a type T, which must provide an
   * operator(), returning a bool.
@@ -30,33 +30,24 @@ template <typename T>
 class CCheckQueue
 {
 private:
-    //! Mutex to protect the inner state
+    //! Mutex to ensure that sleeping threads are woken.
     boost::mutex mutex;
 
     //! Worker threads block on this when out of work
     boost::condition_variable condWorker;
 
-    //! Master thread blocks on this when out of work
-    boost::condition_variable condMaster;
-
-    //! The queue of elements to be processed.
-    //! As the order of booleans doesn't matter, it is used as a LIFO (stack)
-
-    //! The number of workers (including the master) that are idle.
-    int nIdle;
-
-    //! The total number of workers (including the master).
-    int nTotal;
-
     //! The temporary evaluation result.
-    bool fAllOk;
+    std::atomic<uint8_t> fAllOk;
 
     /**
-     * Number of verifications that haven't completed yet.
-     * This includes elements that are no longer queued, but still in the
-     * worker's own batches.
+     * Number of verification threads that aren't in stand-by. When a thread is
+     * awake it may have a job that will return false, but is yet to report the
+     * result through fAllOk.
      */
-    unsigned int nTodo;
+    std::atomic<unsigned int> nAwake;
+
+    /** If there is presently a master process either in the queue or adding jobs */
+    std::atomic<bool> fMasterPresent;
 
     //! Whether we're shutting down.
     bool fQuit;
@@ -64,61 +55,79 @@ private:
     //! The maximum number of elements to be processed in one batch
     unsigned int nBatchSize;
 
+    //! A pointer to contiguous memory that contains all checks
+    T* check_mem {nullptr};
+
+    /** The begin and end offsets into check_mem. 128 bytes of padding is
+     * inserted before and after check_mem_top to eliminate false sharing*/
+    std::atomic<uint32_t> check_mem_bot {0};
+    unsigned char _padding[128];
+    std::atomic<uint32_t> check_mem_top {0};
+    unsigned char _padding2[128];
+
     /** Internal function that does bulk of the verification work. */
     bool Loop(bool fMaster = false)
     {
-        boost::condition_variable& cond = fMaster ? condMaster : condWorker;
-        typename std::vector<T>::iterator checks_iterator;
-        unsigned int nNow = 0;
-        bool fOk = true;
-        do {
+
+        uint8_t fOk = 1;
+        // first iteration, only count non-master threads
+        if (!fMaster)
+            ++nAwake;
+        for (;;) {
+            uint32_t top_cache = check_mem_top;
+            uint32_t bottom_cache = check_mem_bot;
+            // Try to increment bottom_cache by 1 if our version of bottom_cache
+            // indicates that there is work to be done.
+            // E.g., if bottom_cache = top_cache, don't attempt to exchange.
+            //       if  bottom_cache < top_cache, then do attempt to exchange 
+            //
+            // compare_exchange_weak, on failure, updates bottom_cache to latest
+            while (top_cache > bottom_cache && 
+                    !check_mem_bot.compare_exchange_weak( bottom_cache, bottom_cache+1));
+            // If our loop terminated because of no_work_left...
+            if (top_cache <= bottom_cache)
             {
-                boost::unique_lock<boost::mutex> lock(mutex);
-                // first do the clean-up of the previous loop run (allowing us to do it in the same critsect)
-                if (nNow) {
-                    fAllOk &= fOk;
-                    nTodo -= nNow;
-                    if (nTodo == 0 && !fMaster)
-                        // We processed the last element; inform the master it can exit and return the result
-                        condMaster.notify_one();
-                } else {
-                    // first iteration
-                    nTotal++;
-                }
-                // logically, the do loop starts here
-                while (check_mem_top == check_mem_bottom) { // while (empty)
-                    if ((fMaster || fQuit) && nTodo == 0) {
-                        nTotal--;
-                        bool fRet = fAllOk;
-                        // reset the status for new work later
-                        if (fMaster)
-                            fAllOk = true;
-                        // return the current status
-                        return fRet;
+                if (fMaster) {
+                    fMasterPresent = false;
+                    // There's no harm to the master holding the lock
+                    // at this point because all the jobs are taken.
+                    // so busy spin until no one else is awake
+                    while (nAwake) {}
+                    bool fRet = fAllOk;
+                    // reset the status for new work later
+                    fAllOk = 1;
+                    // return the current status
+                    return fRet;
+                } else  if (!fMasterPresent) { // Read once outside the lock and once inside
+                    --nAwake; 
+                    // Unfortunately we need this lock for this to be safe
+                    // We hold it for the min time possible
+                    {
+                        boost::unique_lock<boost::mutex> lock(mutex);
+                        condWorker.wait(lock, [&]{ return fMasterPresent.load();});
                     }
-                    nIdle++;
-                    cond.wait(lock); // wait
-                    nIdle--;
+                    ++nAwake;
                 }
-                // Decide how many work units to process now.
-                // * Do not try to do everything at once, but aim for increasingly smaller batches so
-                //   all workers finish approximately simultaneously.
-                // * Try to account for idle jobs which will instantly start helping.
-                // * Don't do batches smaller than 1 (duh), or larger than nBatchSize.
-                nNow = std::max(1U, std::min(nBatchSize, (unsigned int)(check_mem_top - check_mem_bottom) / (nTotal + nIdle + 1)));
-                checks_iterator = check_mem_bottom;
-                std::advance(check_mem_bottom, nNow);
-                // Check whether we need to do work at all
+            } else {
+                // We compute using bottom_cache (not bottom_cache + 1 as above) because it is 0-indexed
+                T * pT = check_mem + bottom_cache;
+                // Check whether we need to do work at all (can be read outside
+                // of lock because it's fine if a worker executes checks
+                // anyways)
                 fOk = fAllOk;
-            }
-            // execute work
-            for (unsigned int i = 0; i < nNow && fOk; i++) {
-                fOk = (*checks_iterator)();
+                // execute work
+                fOk &= fOk && (*pT)();
+                // We swap in a default constructed value onto pT before freeing
+                // so that we don't accidentally double free when check_mem is
+                // freed. We don't strictly need to free here, but it's good
+                // practice in case T uses a lot of memory.
                 auto t = T();
-                checks_iterator->swap(t);
-                std::advance(checks_iterator, 1);
+                pT->swap(t);
+                // Can't reveal result until after swapped, otherwise
+                // the master thread might exit and we'd double free pT
+                fAllOk &= fOk;
             }
-        } while (true);
+        }
     }
 
 public:
@@ -126,7 +135,7 @@ public:
     boost::mutex ControlMutex;
 
     //! Create a new check queue
-    CCheckQueue(unsigned int nBatchSizeIn) : nIdle(0), nTotal(0), fAllOk(true), nTodo(0), fQuit(false), nBatchSize(nBatchSizeIn) {}
+    CCheckQueue(unsigned int nBatchSizeIn) :  fAllOk(1), nAwake(0), fMasterPresent(false),  fQuit(false), nBatchSize(nBatchSizeIn), check_mem(nullptr), check_mem_bot(0), check_mem_top(0) {}
 
     //! Worker thread
     void Thread()
@@ -140,26 +149,22 @@ public:
         return Loop(true);
     }
 
-    typename std::vector<T>::iterator check_mem;
-    typename std::vector<T>::iterator check_mem_top;
-    typename std::vector<T>::iterator check_mem_bottom;
-    void Setup(typename std::vector<T>::iterator check_mem_in) 
+    //! Setup is called once per batch to point the CheckQueue to the
+    // checks & restart the counters.
+    void Setup(T* check_mem_in)
     {
-        boost::unique_lock<boost::mutex> lock(mutex);
         check_mem = check_mem_in;
-        check_mem_top = check_mem_in;
-        check_mem_bottom = check_mem_in;
+        check_mem_top = 0;
+        check_mem_bot = 0;
+        fMasterPresent = true;
+        boost::unique_lock<boost::mutex> lock(mutex);
+        condWorker.notify_all();
     }
+
     //! Add a batch of checks to the queue
     void Add(size_t size)
     {
-        boost::unique_lock<boost::mutex> lock(mutex);
         check_mem_top += size;
-        nTodo += size;
-        if (size == 1)
-            condWorker.notify_one();
-        else if (size > 1)
-            condWorker.notify_all();
     }
 
     ~CCheckQueue()
@@ -168,7 +173,7 @@ public:
 
 };
 
-/** 
+/**
  * RAII-style controller object for a CCheckQueue that guarantees the passed
  * queue is finished before continuing.
  */
@@ -190,7 +195,7 @@ public:
         if (pqueue != NULL) {
             ENTER_CRITICAL_SECTION(pqueue->ControlMutex);
             check_mem.reserve(size);
-            pqueue->Setup(check_mem.begin());
+            pqueue->Setup(check_mem.data());
         }
     }
 
@@ -203,6 +208,8 @@ public:
         return fRet;
     }
 
+    //! Deprecated. emplacement Add + Flush are the preferred method for adding
+    // checks to the queue.
     void Add(std::vector<T>& vChecks)
     {
         if (pqueue != NULL) {
@@ -214,6 +221,10 @@ public:
             pqueue->Add(s);
         }
     }
+
+    //! Add directly constructs a check on the Controller's memory
+    // Checks created via emplacement Add won't be executed
+    // until a subsequent Flush call.
     template<typename ... Args>
     void Add(Args && ... args)
     {
@@ -221,6 +232,8 @@ public:
             check_mem.emplace_back(std::forward<Args>(args)...);
         }
     }
+
+    //! FLush is called to inform the worker of new jobs
     void Flush(size_t s)
     {
         if (pqueue != NULL) {