--- a/src/checkqueue.h
+++ b/src/checkqueue.h
@@ -41,7 +41,6 @@ private:
 
     //! The queue of elements to be processed.
     //! As the order of booleans doesn't matter, it is used as a LIFO (stack)
-    std::vector<T> queue;
 
     //! The number of workers (including the master) that are idle.
     int nIdle;
@@ -69,8 +68,7 @@ private:
     bool Loop(bool fMaster = false)
     {
         boost::condition_variable& cond = fMaster ? condMaster : condWorker;
-        std::vector<T> vChecks;
-        vChecks.reserve(nBatchSize);
+        typename std::vector<T>::iterator checks_iterator;
         unsigned int nNow = 0;
         bool fOk = true;
         do {
@@ -88,7 +86,7 @@ private:
                     nTotal++;
                 }
                 // logically, the do loop starts here
-                while (queue.empty()) {
+                while (check_mem_top == check_mem_bottom) { // while (empty)
                     if ((fMaster || fQuit) && nTodo == 0) {
                         nTotal--;
                         bool fRet = fAllOk;
@@ -107,22 +105,19 @@ private:
                 //   all workers finish approximately simultaneously.
                 // * Try to account for idle jobs which will instantly start helping.
                 // * Don't do batches smaller than 1 (duh), or larger than nBatchSize.
-                nNow = std::max(1U, std::min(nBatchSize, (unsigned int)queue.size() / (nTotal + nIdle + 1)));
-                vChecks.resize(nNow);
-                for (unsigned int i = 0; i < nNow; i++) {
-                    // We want the lock on the mutex to be as short as possible, so swap jobs from the global
-                    // queue to the local batch vector instead of copying.
-                    vChecks[i].swap(queue.back());
-                    queue.pop_back();
-                }
+                nNow = std::max(1U, std::min(nBatchSize, (unsigned int)(check_mem_top - check_mem_bottom) / (nTotal + nIdle + 1)));
+                checks_iterator = check_mem_bottom;
+                std::advance(check_mem_bottom, nNow);
                 // Check whether we need to do work at all
                 fOk = fAllOk;
             }
             // execute work
-            BOOST_FOREACH (T& check, vChecks)
-                if (fOk)
-                    fOk = check();
-            vChecks.clear();
+            for (unsigned int i = 0; i < nNow && fOk; i++) {
+                fOk = (*checks_iterator)();
+                auto t = T();
+                checks_iterator->swap(t);
+                std::advance(checks_iterator, 1);
+            }
         } while (true);
     }
 
@@ -145,18 +140,25 @@ public:
         return Loop(true);
     }
 
-    //! Add a batch of checks to the queue
-    void Add(std::vector<T>& vChecks)
+    typename std::vector<T>::iterator check_mem;
+    typename std::vector<T>::iterator check_mem_top;
+    typename std::vector<T>::iterator check_mem_bottom;
+    void Setup(typename std::vector<T>::iterator check_mem_in) 
     {
         boost::unique_lock<boost::mutex> lock(mutex);
-        BOOST_FOREACH (T& check, vChecks) {
-            queue.push_back(T());
-            check.swap(queue.back());
-        }
-        nTodo += vChecks.size();
-        if (vChecks.size() == 1)
+        check_mem = check_mem_in;
+        check_mem_top = check_mem_in;
+        check_mem_bottom = check_mem_in;
+    }
+    //! Add a batch of checks to the queue
+    void Add(size_t size)
+    {
+        boost::unique_lock<boost::mutex> lock(mutex);
+        check_mem_top += size;
+        nTodo += size;
+        if (size == 1)
             condWorker.notify_one();
-        else if (vChecks.size() > 1)
+        else if (size > 1)
             condWorker.notify_all();
     }
 
@@ -174,18 +176,21 @@ template <typename T>
 class CCheckQueueControl
 {
 private:
-    CCheckQueue<T> * const pqueue;
+    std::vector<T> check_mem;
+    CCheckQueue<T>* pqueue;
     bool fDone;
 
 public:
     CCheckQueueControl() = delete;
     CCheckQueueControl(const CCheckQueueControl&) = delete;
     CCheckQueueControl& operator=(const CCheckQueueControl&) = delete;
-    explicit CCheckQueueControl(CCheckQueue<T> * const pqueueIn, const unsigned int size) : pqueue(pqueueIn), fDone(false)
+    explicit CCheckQueueControl(CCheckQueue<T> * const pqueueIn, const unsigned int size) : check_mem(), pqueue(pqueueIn), fDone(false)
     {
         // passed queue is supposed to be unused, or NULL
         if (pqueue != NULL) {
             ENTER_CRITICAL_SECTION(pqueue->ControlMutex);
+            check_mem.reserve(size);
+            pqueue->Setup(check_mem.begin());
         }
     }
 
@@ -200,8 +205,27 @@ public:
 
     void Add(std::vector<T>& vChecks)
     {
-        if (pqueue != NULL)
-            pqueue->Add(vChecks);
+        if (pqueue != NULL) {
+            auto s = vChecks.size();
+            for (T& x : vChecks) {
+                check_mem.emplace_back();
+                check_mem.back().swap(x);
+            }
+            pqueue->Add(s);
+        }
+    }
+    template<typename ... Args>
+    void Add(Args && ... args)
+    {
+        if (pqueue != NULL) {
+            check_mem.emplace_back(std::forward<Args>(args)...);
+        }
+    }
+    void Flush(size_t s)
+    {
+        if (pqueue != NULL) {
+            pqueue->Add(s);
+        }
     }
 
     ~CCheckQueueControl()