--- a/src/validation.h
+++ b/src/validation.h
@@ -45,6 +45,9 @@ class CValidationInterface;
 class CValidationState;
 struct ChainTxData;
 
+template <typename T>
+class CCheckQueueControl;
+
 struct PrecomputedTransactionData;
 struct LockPoints;
 
@@ -378,7 +381,7 @@ int64_t GetTransactionSigOpCost(const CTransaction& tx, const CCoinsViewCache& i
  * instead of being performed inline.
  */
 bool CheckInputs(const CTransaction& tx, CValidationState &state, const CCoinsViewCache &view, bool fScriptChecks,
-                 unsigned int flags, bool cacheStore, PrecomputedTransactionData& txdata, std::vector<CScriptCheck> *pvChecks = NULL);
+                 unsigned int flags, bool cacheStore, PrecomputedTransactionData& txdata, CCheckQueueControl<CScriptCheck> *pCheckQueueControl = NULL);
 
 /** Apply the effects of this transaction on the UTXO set represented by view */
 void UpdateCoins(const CTransaction& tx, CCoinsViewCache& inputs, int nHeight);