--- a/src/validation.h
+++ b/src/validation.h
@@ -17,6 +17,7 @@
 #include "script/script_error.h"
 #include "sync.h"
 #include "versionbits.h"
+#include "consensus/consensus.h"
 
 #include <algorithm>
 #include <exception>
@@ -149,6 +150,16 @@ struct BlockHasher
 {
     size_t operator()(const uint256& hash) const { return hash.GetCheapHash(); }
 };
+
+
+/** The minimum serialized size of a CTxIn even with an empty scriptSig */
+static const unsigned int MIN_TXIN_SERIALIZED_SIZE = 41;
+/** The maximum number of scriptchecks which an input can generate */
+static const unsigned int MAX_SCRIPTCHECKS_PER_TXIN = 1;
+/** The maximum number of possible inputs included a block */
+static const unsigned int MAX_TXINS_PER_BLOCK = MAX_BLOCK_BASE_SIZE/MIN_TXIN_SERIALIZED_SIZE;
+/** The maximum number of scriptchecks which could be created */
+static const unsigned int MAX_SCRIPTCHECKS_PER_BLOCK = MAX_TXINS_PER_BLOCK * MAX_SCRIPTCHECKS_PER_TXIN;
 
 extern CScript COINBASE_FLAGS;
 extern CCriticalSection cs_main;