@JeremyRubin 2018-03-06
Going to close this for now; it is a good target for future work if someone wants to take another crack at it.

I have a number of further relaxations/tweaks (which need more review) to this PR available here https://github.com/JeremyRubin/bitcoin/pull/1 if anyone is taking it over, they should review those first.